fix(audio): 修复音频选择profile错误的问题
根据文档，profile available 的值只有零 和 非零：

Is this profile available? If this is zero, meaning "unavailable", then it makes no sense to try to activate this profile.

If this is non-zero, it's still not a guarantee that activating the
profile will result in anything useful, it just means that the server
isn't aware of any reason why the profile would definitely be useless.

Log:
Bug: https://pms.uniontech.com/task-view-98970.html
Influence: 无
Change-Id: Ia5071e9ba1cfeb46b9a2fc375b6b069c46b69d31